Refactor workflow instance and data converter
