+++ temporalio/worker/__init__.py
@@ -5,17 +5,47 @@ from .interceptor import (
     ActivityInboundInterceptor,
     ActivityOutboundInterceptor,
     ExecuteActivityInput,
+    ExecuteWorkflowInput,
+    HandleQueryInput,
+    HandleSignalInput,
     Interceptor,
+    StartActivityInput,
+    StartChildWorkflowInput,
+    StartLocalActivityInput,
+    WorkflowInboundInterceptor,
 )
 from .worker import Worker, WorkerConfig
+from .workflow_instance import (
+    UnsandboxedWorkflowRunner,
+    WorkflowInstance,
+    WorkflowInstanceDetails,
+    WorkflowRunner,
+)
 
 __all__ = [
+    # Primary types
     "Worker",
     "WorkerConfig",
-    "ExecuteActivityInput",
+    # Interceptor base classes
     "Interceptor",
     "ActivityInboundInterceptor",
     "ActivityOutboundInterceptor",
+    "WorkflowInboundInterceptor",
+    "WorkflowOutboundInterceptor",
+    # Interceptor input
+    "ExecuteActivityInput",
+    "ExecuteWorkflowInput",
+    "HandleSignalInput",
+    "HandleQueryInput",
+    "StartActivityInput",
+    "StartChildWorkflowInput",
+    "StartLocalActivityInput",
+    # Advanced activity classes
     "SharedStateManager",
     "SharedHeartbeatSender",
+    # Advanced workflow classes
+    "WorkflowRunner",
+    "WorkflowInstance",
+    "WorkflowInstanceDetails",
+    "UnsandboxedWorkflowRunner",
 ]
